feat: add nuxt-auth0-github submodule
AI: I've analyzed the diff and created a commit message following the conventional commit format. The diff shows the addition of a Git submodule named "nuxt-auth0-github" pointing to a GitHub repository, as well as some changes to public/admin/index.html where configuration content was removed. The primary change is the addition of the submodule, which I've highlighted in the commit message.